### 2021-04-20 #3196 @whedon
[PRE REVIEW]: S(ound)lab: A teaching-oriented Python package for designing and running psychoacoustic experiments.
**Submitting author:** @DrMarc (<a href="http://orcid.org/0000-0002-2023-1207">Marc Schönwiesner</a>)
**Repository:** <a href="https://github.com/DrMarc/slab" target ="_blank">https://github.com/DrMarc/slab</a>
**Version:** v0.9.0
