### 2021-04-20 4b57cda
dont print every iteration in Trialsequence.transitions
