### 2021-04-20 16e13e1
some small fixes in psychoacoustics docs
