+++ docs/psychoacoustics.rst
@@ -28,7 +28,7 @@ For an example, lets generate pure tones with different frequencies and play the
 
 Usually, we do not only want to play sounds to the subjects in our experiment. Instead, we want them to perform some
 kind of task and give a response. In the example above we could for instance ask after every tone if that tone
-was higher or lower in frequency than the previous one. The response is captured with the :meth:`~slab.psychoacoustics.Key`
+was higher or lower in frequency than the previous one. The response is captured with the :meth:`~slab.psychoacoustics.key`
 context manager which can record single button presses using the :mod:`curses` module. In our example, we instruct the
 subject to press "y" (yes) if the played tone was higher then the previous and "n" (no) if it was lower. After each
 trial we check if the response was correct and store that information as 1 (correct) or 0 (wrong) in the trial sequence.::
@@ -37,7 +37,7 @@ trial we check if the response was correct and store that information as 1 (corr
     stimulus = slab.Sound.tone(frequency=freq)
     stimulus.play()
     if seq.this_n > 0:  # don't get response for first trial
-      previous = seq.conditions[seq.trials[seq.this_n-1]-1]
+      previous = seq.get_future_trial(-1)
       with slab.key() as key:  # wait for a key press
         response = key.getch()
       # check if the response was correct, if so store a 1, else store 0
@@ -233,8 +233,14 @@ As you can see, even through the threshold in the response simulation is 3 (that
 Recording responses
 ^^^^^^^^^^^^^^^^^^^
 When you use a staircase in a listening experiment, you need to record responses from the participant,
-usually in the form of button presses. The :meth:`~slab.psychoacoustics.Key` context manager can record single button presses
-from the computer keyboard (or an attached number pad) using the :mod:`curses` module, or from a custom USB buttonbox. The input is selected by setting :attr:`slab.psychoacoustics.input_method` to 'keyboard' or 'buttonbox'. This allow you to test your code on your laptop and switch to button box input at the lab computer by changing a single line of code. Getting a button press from the keyboard will clear your terminal while waiting for the response, and restore it afterwards. Here is an example of how to use the function in a staircase that finds the detection threshold for a 500 Hz tone:
+usually in the form of button presses. The :meth:`~slab.psychoacoustics.key` context manager can record single button presses
+from the computer keyboard (or an attached number pad) using the :mod:`curses` module, or from a custom USB buttonbox.
+The input is selected by setting :attr:`slab.psychoacoustics.input_method` to 'keyboard' or 'buttonbox'.
+This allow you to test your code on your laptop and switch to button box input at the lab computer by changing a
+single line of code. Getting a button press from the keyboard will clear your terminal while waiting for the response,
+and restore it afterwards. Here is an example of how to use the function in a staircase that finds the
+detection threshold for a 500 Hz tone, after every trial you have to indicate whether you could or could not hear the
+sound py pressing "y" for yes or any other button for no:
 
 .. _detection_example:
 
@@ -245,10 +251,13 @@ from the computer keyboard (or an attached number pad) using the :mod:`curses` m
     for level in stairs:
         stimulus.level = level
         stimulus.play()
-        with slab.Key() as key:
+        with slab.key() as key:
             response = key.getch()
-        stairs.add_response(response) # initiates calculation of next stimulus value
-        stairs.plot()
+        if response == 121:  # 121 is the unicode for the "y" key
+            stairs.add_response(True) # initiates calculation of next stimulus value
+        else:
+            stairs.add_response(False)
+    stairs.plot()
     stairs.threshold()
 
[…]
