### 2021-04-20 0764d29
some minor fixes, mostly for psychoacoustics documentation (#19)
* can you use apt-get in workflow?

* use sudo in workflow?

* test workflow

* install pulseaudio

* test workflow

* test workflow

* test workflow

* test workflow

* add pyinstaller

* remove pyinstaller

* use apt clean before installing pulseaudio

* test workflow

* exclude dynamicripple from pytest

* test workflow

* add flake8 config file

* exclude soundfile from flake8

* exclude soundcard from flake8

* ignore soundcard for pytest

* ignore soundfile for pytest

* fix bu in test routine

* don't check sound level

* 
[…]
