### 2021-04-21 #3196 @arfon
:wave: @DrMarc - thanks for your submission to JOSS. A few housekeeping issues here related to your paper:

- Please either add an ORCID for `Ole Bialas` or remove the blank entry for them.
- Please check the suggested DOI above from @whedon, and if it looks correct, then please update your BibTeX file with the corresponding DOI.

Also, could you confirm the main purpose of this software? From my initial read of your paper, this sounds like it's primary purpose is as a teaching tool, which is not in scope for JOSS (but is in scope for our sister journal JOSE, [_Journal of Open Source Education_](http://jose.theoj.org)).
