### 2021-04-22 #3196 @DrMarc
Hi,
thanks for your comment. The package is aimed at designing and running actual experiments and is used in that capacity in our lab (several papers using it are forthcoming). Since we have many Master and Bachelor students, we also wanted something that will allow them to go from zero Python experience to actual research within the typical 6-months timespan of a thesis. This turned out to be very useful in our lab and resulted in a productivity boost, which is why we discuss it prominently in the paper. If you think it's overdone or sounds out of scope, we can tone that aspect down.
Cheers,
Marc
