### 2021-04-26 f57e54c
catch h5py warning when loading sofa file
