+++ paper/paper.md
@@ -1,5 +1,5 @@
 ---
-title: 's(ound)lab: A teaching-oriented Python package for designing and running psychoacoustic experiments.'
+title: 's(ound)lab: An easy to learn Python package for designing and running psychoacoustic experiments.'
 tags:
   - Python
   - psychoacoustics
@@ -23,9 +23,7 @@ bibliography: paper.bib
 
 ---
 # Summary
-Most life science undergraduate students join our lab without prior training in computer programming and digital signal processing.
-The primary aim of slab is to enable these students to learn Python, implement novel psychoacoustic experiments, and ultimately complete their theses on time.
-To enable quick implementation of experiments, slab implements many of the procedures for psychoacoustic research and experiment control and is easily combined with other Python software. To encourage Python learning, slab provides building blocks rather than ready-made solutions, so that students still need to carefully consider stimulation, sequencing and data management. This also makes slab very flexible and easy to customise. In the documentation (see soundlab.readthedocs.io), we provide tutorials suitable for new students. We also provide actual experiments conducted in our lab using slab as worked examples.
+Slab enables researchers and students to prototype and implement psychoacoustic experiments quickly. Slab implements many of the procedures for psychoacoustic research and experiment control and is easily combined with other Python software. A secondary aim of slab is to enable researchers and students without prior training in computer programming and digital signal processing to implement and conduct these experiments. Slab provides building blocks rather than ready-made solutions, so that experimenters still need to carefully consider stimulation, sequencing and data management. This also makes slab very flexible and easy to customise. In the documentation (see soundlab.readthedocs.io), we provide tutorials suitable for beginners. We also provide experiments conducted in our lab as worked examples.
 
 Slab can:
 
@@ -57,9 +55,9 @@ print(stairs.threshold(n=14))
 ```
 
 # Statement of need
-Slab was written to address our own need for a Python package that allows incoming students to implement their own experiments with clean and maintainable code. Students and researchers should be able to write and understand the code that they are using. Several students have now learned Python and completed their theses using slab, and we think the package may be useful to others in the same situation. Our approach differs from existing software packages for running behavioural experiments, which provide a high level graphical user interface to customise the parameters of pre-made experiments [@psychopy2; @pychoacoustics]. In our experience, this leads to very little generalisable learning of Python and experimental control. Slab facilitates this learning by providing basic building blocks, implemented concisely in pure Python, that can be used to construct experiments of various levels of complexity.
+Slab was written to address our own need for a Python package that allows incoming students and new researchers to implement their own experiments with clean and maintainable code. Experimenters should be able to write and understand the code that they are using. Many students in our lab have very quickly progressed from programming novices to completing psychoacoustic research theses using slab, and we think the package may be useful to others in the same situation. Our approach differs from existing software packages for running behavioural experiments, which provide a high level graphical user interface to customise the parameters of pre-made experiments [@psychopy2; @pychoacoustics]. In our experience, this leads to very little generalisable learning of Python and experimental control. Slab facilitates this learning by providing basic building blocks, implemented concisely in pure Python, that can be used to construct experiments of various levels of complexity.
 There is some overlap with librosa [@mcfee2015librosa], a Python package for music analysis, but that package focusses on feature extraction and does not support psychoacoustic experimentation.
-Slab is also one of very few that features manipulation of head-related transfer functions and a simple API for reading a standard file format (SOFA) for such data. There is overlap with more recent implementations of the complete SOFA API [@pysofaconventions; @python-sofa], but these packages provide no methods for typical experimental manipulations of head-related transfer functions. We will likely use `pysofaconventions` internally for handling SOFA files within `slab` in the near future.
[…]
