+++ paper/paper.md
@@ -56,7 +56,7 @@ print(stairs.threshold(n=14))
 
 # Statement of need
 Slab was written to address our own need for a Python package that allows incoming students and new researchers to implement their own experiments with clean and maintainable code. Experimenters should be able to write and understand the code that they are using. Many students in our lab have very quickly progressed from programming novices to completing psychoacoustic research theses using slab, and we think the package may be useful to others in the same situation. Our approach differs from existing software packages for running behavioural experiments, which provide a high level graphical user interface to customise the parameters of pre-made experiments [@psychopy2; @pychoacoustics]. In our experience, this leads to very little generalisable learning of Python and experimental control. Slab facilitates this learning by providing basic building blocks, implemented concisely in pure Python, that can be used to construct experiments of various levels of complexity.
-There is some overlap with librosa [@mcfee2015librosa], a Python package for music analysis, but that package focusses on feature extraction and does not support psychoacoustic experimentation.
+There is some overlap with librosa [@mcfee2015librosa], a Python package for music analysis, but that package focuses on feature extraction and does not support psychoacoustic experimentation.
 Slab is also one of very few packages that features manipulation of head-related transfer functions and a simple API for reading a standard file format (SOFA) for such data. There is overlap with more recent implementations of the complete SOFA API [@pysofaconventions; @python-sofa], but these packages provide no methods for typical experimental manipulations of head-related transfer functions. We will likely use `pysofaconventions` internally for handling SOFA files within `slab` in the near future.
 The architecture of the `Signal` class and some of the sound generation methods in the `Sound` class are inspired on the 1.4 version of Brian.hears [@brian2hears], but we made several simplifications based on learning reports from students. For instance, signal objects do not implement buffering and do not inherit from Numpy arrays [@harris2020array] directly, because these features significantly hindered students' understanding of the code.
 
