### 2021-04-28 #3196 @arfon
> @arfon are you happy with the above ☝️ comments on the applications of this work?

:+1: yep, this sounds reasonable to me. I'm also happy to take this submission on.
