### 2021-05-02 #3196 @DrMarc
Hi @arfon ,
Many thanks, glad to hear it. Below are the reviewers from the list that have listed relevant expertise in audio signal processing and/or psychophysics related to sound or speech stimuli:
peircej, jonathan.peirce@nottingham.ac.uk
trettenbrein, trettenbrein@cbs.mpg.de
JonathanReardon, jonathanjreardon@gmail.com
sneakers-the-rat, jlsaunders987@gmail.com
faroit, mail@faroit.com
hadware, hadrien.titeux@ens.fr

Cheers,
Marc
