### 2021-05-06 956c99f
change docstrings for proper sphinx rendering
