+++ slab/hrtf.py
@@ -31,7 +31,9 @@ from slab.filter import Filter
 
 
 class HRTF:
-    """ Class for reading and manipulating head-related transfer functions with attributes and functions to manage them.
+    """
+    Class for reading and manipulating head-related transfer functions with attributes and functions to manage them.
+
     Arguments:
         data (str | Filter | numpy.ndarray): Typically, this is the path to a file in the .sofa format.
             The file is then loaded and the data of each source for which the transfer function was recorded is stored
@@ -52,12 +54,14 @@ class HRTF:
         .listener (dict): a dictionary containing the position of the listener ("pos"), the point which the listener
             is fixating ("view"), the point 90° above the listener ("up") and vectors from the listener to those points.
         .samplerate (float): sampling rate at which the HRTF data was acquired.
-    Example:
+    Example::
+
         from slab import data_path, HRTF
         hrtf = slab.HRTF(data=data_path()+'mit_kemar_normal_pinna.sofa')  # initialize from sofa file
         sourceidx = hrtf.cone_sources(20)
         hrtf.plot_sources(sourceidx)
-        hrtf.plot_tf(sourceidx, ear='left') """
+        hrtf.plot_tf(sourceidx, ear='left')
+    """
     # instance properties
     n_sources = property(fget=lambda self: len(self.sources),
                          doc='The number of sources in the HRTF.')
@@ -117,11 +121,14 @@ class HRTF:
     # Static methods (used in __init__)
     @staticmethod
     def _sofa_load(filename, verbose=False):
-        """ Read a SOFA file.
+        """
+        Read a SOFA file.
+
         Arguments:
             filename (str): full path to the .sofa file.
         Returns:
-            (h5netcdf.core.File): the data from the .sofa file. """
+            (h5netcdf.core.File): the data from the .sofa file.
+        """
         if h5netcdf is False:
             raise ImportError('Reading from sofa files requires h5py and h5netcdf.')
         with warnings.catch_warnings():
@@ -133,8 +140,10 @@ class HRTF:
 
     @staticmethod
     def _sofa_get_samplerate(f):
-        """ Returns the sampling rate of the recordings. If the sampling rate is not given in Hz, the function assumes
+        """
+        Returns the sampling rate of the recordings. If the sampling rate is not given in Hz, the function assumes
         it is given in kHz and multiplies by 1000 to convert to Hz.
+
         Arguments:
             f (h5netcdf.core.File): data as returned by the `_sofa_load` method.
         Returns:
@@ -149,7 +158,9 @@ class HRTF:
 
     @staticmethod
     def _sofa_get_sourcepositions(f):
-        """ Returns an array of positions of all sound sources.
+        """
+        Returns an array of positions of all sound sources.
+
         Arguments:
             f (h5netcdf.core.File): data as returned by the _sofa_load method.
         Returns:
@@ -174,12 +185,15 @@ class HRTF:
 
     @staticmethod
     def _sofa_get_listener(f):
-        """ Returns dict with listeners positional information - used for plotting.
+        """
+        Returns dict with listeners positional information - used for plotting.
+
         Attributes:
             f (h5netcdf.core.File): data as returned by the `_sofa_load()` method.
         Returns:
             (dict): position of the listener ("pos"), the point which the listener is fixating ("view")
-                the point 90° above the listener ("up") and vectors from the listener to those points. """
+                the point 90° above the listener ("up") and vectors from the listener to those points.
+        """
         lis = {'pos': numpy.array(f.variables['ListenerPosition'], dtype='float')[0],
                'view': numpy.array(f.variables['ListenerView'], dtype='float')[0],
[…]
