+++ slab/binaural.py
@@ -12,6 +12,7 @@ class Binaural(Sound):
     generation functions  from the Sound class, but returns binaural signals. Recasting an object of class sound or
     sound with 1 or 3+ channels calls Sound.copychannel to return a binaural sound with two channels identical
     to the first channel of the original sound.
+
     Arguments:
         data (slab.Signal | numpy.ndarray | list | str): see documentation of slab.Sound for details. the `data` must
             have either one or two channels. If it has one, that channel is duplicated
@@ -22,7 +23,8 @@ class Binaural(Sound):
         .data: the data-array of the Sound object which has the shape `n_samples` x `n_channels`.
         .n_channels: the number of channels in `data`. Must be 2 for a binaural sound.
         .n_samples: the number of samples in `data`. Equals `duration` * `samplerate`.
-        .duration: the duration of the sound in seconds. Equals `n_samples` / `samplerate`. """
+        .duration: the duration of the sound in seconds. Equals `n_samples` / `samplerate`.
+    """
     # instance properties
     def _set_left(self, other):
         if hasattr(other, 'samplerate'):  # probably an slab object
@@ -71,9 +73,11 @@ class Binaural(Sound):
             ValueError('Binaural sounds must have two channels!')
 
     def itd(self, duration=None, max_lag=0.001):
-        """ Either estimate the interaural time difference of the sound or generate a new sound with the specified
-            interaural time difference. The resolution for computing the ITD is 1/samplerate seconds. A negative
-            ITD value means that the right channel is delayed, meaning the sound source is to the left.
+        """
+        Either estimate the interaural time difference of the sound or generate a new sound with the specified
+        interaural time difference. The resolution for computing the ITD is 1/samplerate seconds. A negative
+        ITD value means that the right channel is delayed, meaning the sound source is to the left.
+
         Arguments:
             duration (None| int | float): Given None, the instance's ITD is computed. Given another value, a new sound
                 with the desired interaural time difference in samples (given an integer) or seconds (given a float)
@@ -83,10 +87,12 @@ class Binaural(Sound):
         Returns:
              (int | slab.Binaural): The interaural time difference in samples or a copy of the instance with the
                 specified interaural time difference.
-        Examples:
+        Examples::
+
             sound = slab.Binaural.whitenoise()
             lateral = sound.itd(duration=0.0005)  # generate a sound with 0.5 ms ITD
-            lateral.itd()  # estimate the ITD of the sound """
+            lateral.itd()  # estimate the ITD of the sound
+        """
         if duration is None:
             return self._get_itd(max_lag)
         else:
@@ -110,18 +116,22 @@ class Binaural(Sound):
         return self.delay(duration=abs(duration), channel=channel)
 
     def ild(self, dB=None):
-        """ Either estimate the interaural level difference of the sound or generate a new sound with the specified
-            interaural level difference. Negative ILD value means that the left channel is louder than the right
-            channel, meaning that the sound source is to the left.
+        """
+        Either estimate the interaural level difference of the sound or generate a new sound with the specified
+        interaural level difference. Negative ILD value means that the left channel is louder than the right
+        channel, meaning that the sound source is to the left.
+
         Arguments:
             dB (None | int | float): If None, estimate the sound's ITD. Given a value, a new sound is generated with
                 the desired interaural level difference in decibels.
         Returns:
             (float | slab.Binaural): The sound's aural level difference or a new instance with the specified ILD.
-        Examples:
[…]

[…]
