### 2021-05-07 #3196 @arfon
👋 @sneakers-the-rat @hadware @peircej - would any of you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html

The submission under review here is [_S(ound)lab: A teaching-oriented Python package for designing and running psychoacoustic experiments_](https://github.com/DrMarc/slab).
