### 2021-05-07 1bb64fa
docs (#20)
* test workflow

* install pulseaudio

* test workflow

* test workflow

* test workflow

* test workflow

* add pyinstaller

* remove pyinstaller

* use apt clean before installing pulseaudio

* test workflow

* exclude dynamicripple from pytest

* test workflow

* add flake8 config file

* exclude soundfile from flake8

* exclude soundcard from flake8

* ignore soundcard for pytest

* ignore soundfile for pytest

* fix bu in test routine

* don't check sound level

* add testing for filters

* change filter documentation

* import re

* change filters to filter in documentation

*
[…]
