+++ slab/psychoacoustics.py
@@ -30,12 +30,15 @@ input_method = 'keyboard'  #: sets the input for the Key context manager to 'key
 
 
 def data_path(allow_download=True):
-    """ Get the path where auxiliary data (like HRTF recordings, calibrations) are stored.
+    """
+    Get the path where auxiliary data (like HRTF recordings, calibrations) are stored.
+
     Arguments:
         allow_download (bool): If true, check if the data folder contains the KEMAR HRTF-recordings and if not,
             download them from the GitHub repository.
     Returns:
-        (str): the path to the data folder. """
+        (str): the path to the data folder.
+    """
     path = str(pathlib.Path(__file__).parent.resolve() / pathlib.Path('data')) + os.sep
     if not os.path.exists(path):
         os.mkdir(path)
@@ -48,10 +51,11 @@ def data_path(allow_download=True):
 
 
 class _Buttonbox:
-    """ Adapter class to allow easy switching between input from the keyboard via curses and from the custom buttonbox
+    """
+    Adapter class to allow easy switching between input from the keyboard via curses and from the custom buttonbox
     adapter (custom arduino device that sends a keystroke followed by a return keystroke when pressing a button on
-    the arduino)."""
-
+    the arduino).
+    """
     @staticmethod
     def getch():
         input_key = input()  # buttonbox adapter has to return the keycode of intended keys!
@@ -60,12 +64,12 @@ class _Buttonbox:
 
 
 class _FigChar:
-    """ Adapter class to allow easy switching to input via the current_character attribute of stairs figure.
+    """
+    Adapter class to allow easy switching to input via the current_character attribute of stairs figure.
     Set slab.psychoacoustics.input_method = 'figure' to use. A figure with the name 'stairs' will be opened if it is not
     already present. If used together with the plot method of the Staircase class, input is acquired through the stairs
     plot. Depending on the operating system, you may have to click once into the figure to give it focus.
     """
-
     import warnings # necessary for matplotlib versions <3.5 to suppress a MatplotlibDeprecationWarning
     import matplotlib.cbook
     warnings.filterwarnings("ignore", category=matplotlib.cbook.MatplotlibDeprecationWarning)
@@ -86,11 +90,15 @@ class _FigChar:
 
 @contextmanager
 def key():
-    """ Wrapper for curses module to simplify getting a single keypress from the terminal (default) or a buttonbox.
+    """
+    Wrapper for curses module to simplify getting a single keypress from the terminal (default) or a buttonbox.
     Set slab.psychoacoustics.input_method = 'buttonbox' to use a custom USB buttonbox.
-    Example:
+
+    Example::
+
         with slab.Key() as key:
-        response = key.getch() """
+        response = key.getch()
+    """
 
     if input_method == 'keyboard':
         if curses is None:
@@ -116,12 +124,15 @@ class LoadSaveMixin:
     """ Mixin to provide loading and saving functions. Supports JSON the pickle format """
 
     def save_pickle(self, file_name, overwrite=False):
-        """ Save the object as pickle file.
+        """
+        Save the object as pickle file.
+
         Arguments:
             file_name (str | pathlib.Path): name of the file to create.
             overwrite (bool): overwrite existing file with the same name, defaults to False.
         Returns:
-            (bool): True if writing was successful, False if not """
+            (bool): True if writing was successful, False if not
+        """
         if isinstance(file_name, pathlib.PosixPath):
             file_name = str(file_name)
         if os.path.isfile(file_name):
@@ -133,9 +144,12 @@ class LoadSaveMixin:
             return True
 
     def load_pickle(self, file_name):
-        """ Read pickle file and deserialize the object into `self.__dict__`.
+        """
[…]
