+++ slab/signal.py
@@ -6,7 +6,7 @@ try:
 except ImportError:
     scipy = False
 
-_default_samplerate = 8000  #: The default samplerate in Hz; used by all methods if on samplerate argument is provided.
+_default_samplerate = 8000  # default samplerate in Hz; used by all methods if on samplerate argument is provided.
 
 
 def set_default_samplerate(samplerate):
@@ -15,8 +15,10 @@ def set_default_samplerate(samplerate):
 
 
 class Signal:
-    """ Base class for Signal data (from which the Sound and Filter class inherit).
+    """
+    Base class for Signal data (from which the Sound and Filter class inherit).
     Provides arithmetic operations, slicing, and conversion between samples and times.
+
     Arguments:
         data (numpy.ndarray | slab.Signal | list): samples of the sound. If it is an array, the first dimension
             should represent the number of samples and the second one the number of channels. If it's an object,
@@ -28,13 +30,15 @@ class Signal:
         .n_samples: duration of the sound in samples
         .n_channels: number of channels in the sound
         .times: list with the time point of each sample
-    Examples:
+    Examples::
+
         import slab, numpy
         sig = slab.Signal(numpy.ones([10,2]),samplerate=10)  # create a sound
         sig[:5] = 0  # set the first 5 samples to 0
         sig[:,1]  # select the data from the second channel
         sig2 = sig * 2  # multiply each sample by 2
-        sig_inv = -sig  # invert the phase """
+        sig_inv = -sig  # invert the phase
+    """
     # instance properties
     n_samples = property(fget=lambda self: self.data.shape[0],
                          doc='The number of samples in the Signal.')
@@ -142,13 +146,16 @@ class Signal:
     # static methods (belong to the class, but can be called without creating an instance)
     @staticmethod
     def in_samples(ctime, samplerate):
-        """ Converts time values in seconds to samples. This is used to enable input in either samples (integers) or
+        """
+        Converts time values in seconds to samples. This is used to enable input in either samples (integers) or
         seconds (floating point numbers) in the class.
+
         Arguments:
             ctime (int | float | numpy | ndarray): the time(s) to convert to samples.
             samplerate (int): the samplerate of the sound.
         Returns:
-             (int | numpy.ndarray): the time(s) in samples. """
+             (int | numpy.ndarray): the time(s) in samples.
+        """
         out = None
         if isinstance(ctime, (int, numpy.int64)):  # single int is treated as samples
             out = ctime
@@ -179,7 +186,9 @@ class Signal:
 
     # instance methods (belong to instances created from the class)
     def channel(self, n):
-        """ Get a single data channel.
+        """
+        Get a single data channel.
+
         Arguments:
             n (int): channel index
         Returns:
@@ -195,11 +204,14 @@ class Signal:
         return (self.channel(i) for i in range(self.n_channels))
 
     def resize(self, duration):
-        """ Change the duration by padding with zeros or cutting the data.
+        """
+        Change the duration by padding with zeros or cutting the data.
+
         Arguments:
             duration (float | int): new duration of the sound in seconds (given a float) or in samples (given an int).
         Returns:
-            (slab.Signal): a new instance of the same class with the specified duration. """
+            (slab.Signal): a new instance of the same class with the specified duration.
+        """
         duration = Signal.in_samples(duration, self.samplerate)
         resized = copy.deepcopy(self)
         if duration == len(self.data):
@@ -212,11 +224,14 @@ class Signal:
         return resized
 
     def resample(self, samplerate):
-        """ Resample the sound.
+        """
+        Resample the sound.
+
         Arguments:
             samplerate (int): the samplerate of the resampled sound.
         Returns:
-            (slab.Signal): a new instance of the same class with the specified samplerate. """
+            (slab.Signal): a new instance of the same class with the specified samplerate.
[…]
