### 2021-05-07 441ece7
added trim method
