### 2021-05-07 4cca3b9
removed duplicate samplerate setting
