### 2021-05-07 2da8fb0
added tests for overloaded sound generators
