### 2021-05-07 7e2fe47
added tests for samplerate and envelope options
