### 2021-05-07 ec67d17
aweight method does not change type, no overloading in Binaural needed
