+++ slab/binaural.py
@@ -507,7 +507,3 @@ class Binaural(Sound):
     def equally_masking_noise(**kwargs):
         """ Identical to slab.Sound.erb_noise, but with two channels. """
         return Binaural(Sound.equally_masking_noise(**kwargs))
-
-    def aweight(self):
-        """ Identical to slab.Sound.aweight, but with two channels. """
-        return Binaural(Sound.aweight(self))
