### 2021-05-07 9f6731a
corrected trim() test
