### 2021-05-07 e853a23
removed deprecated escape seq '\*' in docstrings
