### 2021-05-08 bd91ec0
reduced noise in ils computation, relaxed test
