+++ tests/test_binaural.py
@@ -86,8 +86,7 @@ def test_externalize():
 
 
 def test_interaural_level_spectrum():
-    hrtf = slab.HRTF(slab.data_path() +'mit_kemar_normal_pinna.sofa')
-    ils = slab.Binaural._make_level_spectrum_filter(hrtf)  # provide HRTF to ensure a new filter is made
+    ils = slab.Binaural._make_level_spectrum_filter()
     sound = slab.Binaural.whitenoise(samplerate=int(ils[0, 0]))
     azimuths = ils[0, 1:]
     for i, azi in enumerate(azimuths):
@@ -96,7 +95,7 @@ def test_interaural_level_spectrum():
         fbank = slab.Filter.cos_filterbank(samplerate=sound.samplerate, pass_bands=True)
         subbands_left = fbank.apply(lateral.left)
         subbands_right = fbank.apply(lateral.right)
-        assert -1.5 < (level_differences - (subbands_left.level - subbands_right.level)).mean() < 1.5
+        assert -3 < (level_differences - (subbands_left.level - subbands_right.level)).mean() < 3
 
 def test_overloaded_sound_generators():
     sig = slab.Binaural.pinknoise()
