### 2021-05-08 060f577
cleanup docstrings and comments
