+++ slab/binaural.py
@@ -1,5 +1,9 @@
 import copy
 import numpy
+try:
+    import h5py
+except ImportError:
+    h5py = False
 from slab.sound import Sound
 from slab.signal import Signal
 from slab.filter import Filter
@@ -234,28 +238,28 @@ class Binaural(Sound):
         return itd
 
     @staticmethod
-    def azimuth_to_ild(azimuth, frequency=2000, hrtf=None):
+    def azimuth_to_ild(azimuth, frequency=2000, ils=None):
         """
-        Get the interaural level difference corresponding to a sound source at a given azimuth.
+        Get the interaural level difference corresponding to a sound source at a given azimuth and frequency.
 
         Arguments:
             azimuth (int | float): The azimuth angle of the sound source, negative numbers refer to sources to the left.
             frequency (int | float): Frequency in Hz for which the ITD is estimated.
                 Use the default for for sounds with a broadband spectrum.
-            hrtf (None | slab.HRTF): head-related transfer function from which the ILD is taken.
-                If None use the MIT KEMAR mannequin.
+            ils (None | dict): interaural level spectrum from which the ILD is taken.
+                Generate with `_make_level_spectrum_filter()`. If None, uses level spectrum of the MIT KEMAR mannequin.
         Returns:
             (float): The interaural level difference for a sound source at a given azimuth in decibels.
         Examples::
 
             ild = slab.Binaural.azimuth_to_ild(-90) # ILD equivalent to 90 deg leftward source using KEMAR HRTF.
         """
-        ils = Binaural._make_level_spectrum_filter(hrtf=hrtf)
-        freqs = ils[1:, 0]  # get vector of frequencies in ils filter bank
-        azis = ils[0, 1:]  # get vector of azimuths in ils filter bank
-        ils = ils[1:, 1:]  # the rest is the filter
-        levels = [numpy.interp(azimuth, azis, ils[i, :]) for i in range(ils.shape[0])]  # interpolate levels at azimuth
-        return numpy.interp(frequency, freqs, levels)*-1   # interpolate level difference at frequency
+        if ils is None:
+            ils = Binaural.make_interaural_level_spectrum()
+        # interpolate levels at azimuth:
+        level_diffs = ils['level_diffs']
+        levels = [numpy.interp(azimuth, ils['azimuths'], level_diffs[i, :]) for i in range(level_diffs.shape[0])]
+        return numpy.interp(frequency, ils['frequencies'], levels)*-1   # interpolate level difference at frequency
 
     def at_azimuth(self, azimuth=0):
         """
@@ -297,37 +301,43 @@ class Binaural(Sound):
         return filtered_signal
 
     @staticmethod
-    def _make_level_spectrum_filter(hrtf=None):
+    def make_interaural_level_spectrum(hrtf=None):
         """
         Compute the frequency band specific interaural intensity differences for all sound source azimuth's in
         a head-related transfer function. For every azimuth in the hrtf, the respective transfer function is applied
         to a sound. This sound is then divided into frequency sub-bands. The interaural level spectrum is the level
         difference between right and left for each of these sub-bands for each azimuth.
+        Note that the computation may take a few minutes. When individual HRTFs are not avilable, the level spectrum of
+        the KEMAR mannequin can be used and the level spectrum will in this case be cached in slab.datapath() and loaded
+        when the method is called again (requires the h5py module).
 
         Arguments:
             hrtf (None | slab.HRTF): The head-related transfer function used to compute the level spectrum. If None,
                 use the recordings from the KEMAR mannequin. For the KEMAR, the level spectrum is saved to a file
                 and loaded, the next time this function is executed to save computation time.
         Returns:
[…]
