### 2021-05-10 b6bc222
corrected level spectrum doc string
