### 2021-05-15 #3284 @sneakers-the-rat
hope it's alright if i do my checklist in a separate comment & copy/paste it into the OP at the end so i don't accidentally mess up anything @hadware is doing

The need is clearly described in the manuscript -- a simple script-based psychoacoustics library that incorporates HRTFs. Big fan of modular tools and think that's the route that most scientific software should take, often attempts to, and frequently misses (myself included) so I am interested to see the implementation.
