### 2021-05-15 #3284 @sneakers-the-rat
hope it's alright if i do my checklist in a separate comment & copy/paste it into the OP at the end so i don't accidentally mess up anything @hadware is doing

a naive whole-repo [hercules](https://github.com/src-d/hercules) ownership plot shows that the library is currently roughly a ~50/50 split between the two authors
![slab_burndown](https://user-images.githubusercontent.com/12961499/118346242-d85b2480-b4ee-11eb-9ad8-920c42ce9f28.png)

The need is clearly described in the manuscript -- a simple script-based psychoacoustics library that incorporates HRTFs. Big fan of modular tools and think that's the route that most scientific software should take, often attempts to, and frequently misses (myself included) so I am interested to see the implementation.
