### 2021-05-15 f509c43
added pytest to extras_require and installation instructions for testing
