### 2021-05-16 d0bb937
fixes #27 added _repr_html_ and plot_samples methods
