+++ slab/psychoacoustics.py
@@ -313,8 +313,8 @@ class Trialsequence(collections.abc.Iterator, LoadSaveMixin, TrialPresentationOp
             filename of a previously saved trial sequence object, which is then loaded.
         n_reps (int): number of repetitions for each condition. Number of trials is given by len(conditions)*n_reps).
         trials (None | list | numpy.ndarray): The sequence of trials in the order in which they are appearing in
-            sequence. Defaults to None since the trials are usually generated by the class based on the other
-            parameters. wHowever, it is possible to pass a list or one dimensional array. In that case the parameters
+            sequence. Defaults to None, because trials are usually generated by the class based on the other
+            parameters. However, it is possible to pass a list or one-dimensional array. In that case the parameters
             for generating the sequence are ignored.
         kind (str): The kind of randomization used to generate the trial sequence. Possible options are:
             `non_repeating` (randomization without direct repetition of a condition, default if n_conditions > 2),
@@ -322,7 +322,7 @@ class Trialsequence(collections.abc.Iterator, LoadSaveMixin, TrialPresentationOp
             `infinite` (sequence that reset when reaching the end to generate an infinite number of trials.
             randomization method is random_permutation` if n_conditions` <= 2 and `non_repeating` otherwise).
         deviant_freq (float): frequency with which deviants (encoded as 0) appear in the sequence. The minimum number
-        of trials between two deviants is 3 if deviant frequency is below 10%, 2 if it is below 20% and 1 if it
+            of trials between two deviants is 3 if deviant frequency is below 10%, 2 if it is below 20% and 1 if it
             is below 30%. A deviant frequency greater than 30% is not supported
         label (str): a text label for the sequence.
     Attributes:
@@ -509,7 +509,7 @@ class Trialsequence(collections.abc.Iterator, LoadSaveMixin, TrialPresentationOp
         elif deviant_freq <= .3:
             min_dist = 1
         else:
-            raise ValueError("Deviant frequency can't be greater than 30%!")
+            raise ValueError("Deviant frequency can't be greater than 0.3!")
         # get the possible combinations of deviants and normal trials:
         n_deviants = int(n_standard*deviant_freq)
         indices = range(n_standard)
