### 2021-05-16 a23fa4b
fixes #25, corrected examples
