+++ README.md
@@ -10,13 +10,13 @@
 
 **Slab** ('es-lab', or sound laboratory) is an open source project and Python package that makes working with sounds and running psychoacoustic experiments simple, efficient, and fun! For instance, it takes just eight lines of code to run a pure tone audiogram using an adaptive staircase:
 ```python
-    import slab
-    stimulus = slab.Sound.tone(frequency=500, duration=0.5) # make a 0.5 sec pure tone of 500 Hz
-    stairs = slab.Staircase(start_val=50, n_reversals=10) # set up the adaptive staircase
-    for level in stairs: # the staircase object returns a value between 0 and 50 dB for each trial
-        stimulus.level = level
-        stairs.present_tone_trial(stimulus) # plays the tone and records a keypress (1 for 'heard', 2 for 'not heard')
-    print(stairs.threshold()) # print threshold when done
+import slab
+stimulus = slab.Sound.tone(frequency=500, duration=0.5) # make a 0.5 sec pure tone of 500 Hz
+stairs = slab.Staircase(start_val=50, n_reversals=10) # set up the adaptive staircase
+for level in stairs: # the staircase object returns a value between 0 and 50 dB for each trial
+    stimulus.level = level
+    stairs.present_tone_trial(stimulus) # plays the tone and records a keypress (1 for 'heard', 2 for 'not heard')
+print(stairs.threshold()) # print threshold when done
 ```
 
 Why slab?
@@ -29,77 +29,85 @@ Slab represents sounds as [Numpy](https://www.numpy.org) arrays and provides cla
 
 **Signal**: Provides a generic signal object with properties duration, number of samples, sample times, number of channels. Keeps the data in a 'data' property and implements slicing, arithmetic operations, and conversion between sample points and time points.
 ```python
-sig = slab.Sound.pinknoise(nchannels=2) # make a pink noise
+sig = slab.Sound.pinknoise(n_channels=2) # make a 2-channel pink noise
 sig.duration
-out: 1.0
-sig.nsamples
-out: 8000
+# 1.0
+sig.n_samples
+# 8000
 sig2 = sig.resample(samplerate=4000) # resample to 4 kHz
-env = sig2.envelope() # returns a new sound containing the lowpass Hilbert envelopes of both channels
+env = sig2.envelope() # returns a new signal containing the lowpass Hilbert envelopes of both channels
 sig.delay(duration=0.0006, channel=0) # delay the first channel by 0.6 ms
 ```
 
 **Sound**: Inherits from Signal and provides methods for generating, manipulating, displaying, and analysing sound stimuli. Can compute descriptive sound features and apply manipulations to all sounds in a folder.<sup id="a1">[1](#f1)</sup>
 ```python
 vowel = slab.Sound.vowel(vowel='a', duration=.5) # make a 0.5-second synthetic vowel sound
+vowel.play() # play the sound
 vowel.ramp() # apply default raised-cosine onset and offset ramps
-vowel.filter(kind='bp', f=[50, 3000]) # apply bandpass filter between 50 and 3000 Hz
+vowel.filter(kind='bp', frequency=[50, 3000]) # apply bandpass filter between 50 and 3000 Hz
 vowel.spectrogram() # plot the spectrogram
-vowel.spectrum(low=100, high=4000, log_power=True) # plot a band-limited spectrum
+vowel.spectrum(low_cutoff=100, high_cutoff=4000, log_power=True) # plot a band-limited spectrum
 vowel.waveform(start=0, end=.1) # plot the waveform
 vowel.write('vowel.wav') # save the sound to a WAV file
 vocoded_vowel = vowel.vocode() # run a vocoding algorithm
[…]
