### 2021-05-17 c788cf4
corrected formatting of examples
