+++ docs/hrtf.rst
@@ -1,17 +1,15 @@
-.. _HRTF:
-
-.. note:: The class is at the moment geared towards plotting and analysis of HRTF files in the `sofa format <https://www.sofaconventions.org/>`_, because we needed that functionality for grant applications. The functionality will grow as we start to record and manipulate HRTFs more often.
-.. note:: When we started to writing this code, there was no python module for reading and writing sofa files. Now that `pysofaconventions <https://github.com/andresperezlopez/pysofaconventions>`_ is available, we will at some point switch internally to using that module as backend for reading sofa files, instead of our own limited implementation.
+.. _Hrtfs:
 
 HRTFs
 =====
+
 A head-related transfer function (HRTF) describes the impact of the listeners ears, head and torso on incoming sound
 for every position in space. Knowing the listeners HRTF, you can simulate a sound source at any position by filtering
 it with the transfer function corresponding to that position. The :class:`HRTF` class provides methods for
 manipulating, plotting, and applying head-related transfer functions.
 
-Reading sounds
---------------
+Reading HRTF data
+-----------------
 Typically the :class:`HRTF` class is instantiated by loading a file. The canonical format for HRTF-data is called
 sofa (Spatially Oriented Format for Acoustics). To read sofa files, you need to install the h5netcdf module:
 `pip install h5netcdf`. The module includes a set of standard HRTF recordings from the KEMAR (a mannequin for acoustic
@@ -20,12 +18,15 @@ first time you call this function, the recordings will be downloaded from the so
 calling the :class:`HRTF` class with the name of the file as an argument. Print the resulting object to obtain
 information about the structure of the HRTF data ::
 
-    from slab import data_path
-    hrtf = slab.HRTF(data=data_path()+'mit_kemar_normal_pinna.sofa')
+    hrtf = slab.HRTF(data=slab.data_path()+'mit_kemar_normal_pinna.sofa')
     print(hrtf)
-    Out: <class 'hrtf.HRTF'> sources 710, elevations 14, samples 710, samplerate 44100.0
+    # <class 'hrtf.HRTF'> sources 710, elevations 14, samples 710, samplerate 44100.0
 
 Libraries of many other recordings can be found on the `website of the sofa file format <https://www.sofaconventions.org/>`_.
+
+.. note:: The class is at the moment geared towards plotting and analysis of HRTF files in the `sofa format <https://www.sofaconventions.org/>`_, because we needed that functionality for grant applications. The functionality will grow as we start to record and manipulate HRTFs more often.
+
+.. note:: When we started to writing this code, there was no python module for reading and writing sofa files. Now that `pysofaconventions <https://github.com/andresperezlopez/pysofaconventions>`_ is available, we will at some point switch internally to using that module as backend for reading sofa files, instead of our own limited implementation.
 
 Plotting sources
 --------------------
@@ -37,12 +38,14 @@ highlighted in red. This can be useful when you are selecting source locations f
 that you chose correctly. In the example below, we select sources using the methods :meth:`.elevation_sources`, which
 selects sources along a horizontal sphere slice at a given elevation and :meth:`.cone_sources`, which selects sources
 along a vertical slice through the source sphere in front of the listener a given angular distance away from the
-midline.
+midline:
 
 .. plot::
     :include-source:
-    :context: close-figs
+    :context:
+
     # cone_sources and elevation_sources return lists of indices which are concatenated by adding:
[…]
