### 2021-05-17 5debcaf
enable hierarchical navigation on RTD
