+++ docs/index.rst
@@ -57,8 +57,7 @@ Slab will then use another method to play sounds (winsound on Windows, afplay on
 
 .. toctree::
   :caption: Contents
-  :maxdepth: 2
-  :titlesonly:
+  :maxdepth: 3
 
   introduction
   sounds
