+++ docs/conf.py
@@ -32,7 +32,7 @@ extensions = [
     'sphinx.ext.autodoc',
     'matplotlib.sphinxext.plot_directive',
     'sphinx.ext.doctest',
-]
+    'sphinx_rtd_theme']
 
 plot_pre_code = 'import slab'
 master_doc = 'index'
@@ -42,3 +42,6 @@ intersphinx_mapping = {'python': ('https://docs.python.org/', None),
                        'matplotlib': ('http://matplotlib.org/', None),
                        'numpy': ('http://docs.scipy.org/doc/numpy/', None),
                        'scipy': ('http://docs.scipy.org/doc/scipy/reference/', None)}
+
+html_theme = 'sphinx_rtd_theme'
+html_theme_options = {'collapse_navigation': False}
