+++ docs/index.rst
@@ -57,7 +57,7 @@ Slab will then use another method to play sounds (winsound on Windows, afplay on
 
 .. toctree::
   :caption: Contents
-  :maxdepth: 3
+  :maxdepth: 2
 
   introduction
   sounds
