### 2021-05-17 #3284 @DrMarc
"!>> the individual pages in the docs are relatively long, and you've done a good job of writing headings in your .rst, so it would be a big usability improvement for very cheap to configure the docs to enable hierarchical navigation..."

Good point! We enabled hierarchical navigation as recommended.
