### 2021-05-17 #3284 @DrMarc
"I would strongly recommend a hash-based naming system rather than a hardcoded name"

The hashed file names are an excellent idea. It's implemented now.
