### 2021-05-18 e33de9f
fixes #31, corrected docs string formatting for 'spectral feature'
