### 2021-05-18 024d070
fixes #36 (adding additional install options), fixes #37 (add SoundCard)
