+++ README.md
@@ -142,13 +142,23 @@ Installation
 Install the current stable release from the python package index with pip:
 ```pip install slab```
 
-On Linux, you may need to install libsndfile (required by SoundFile) using your distribution’s package manager, for instance:
+Other requirements
+^^^^^^^^^^^^^^^^^^
+On *Linux*, there is only one requirement outside of Python: you may need to install libsndfile using your distribution’s package manager, for instance:
 
 ```sudo apt-get install libsndfile1```
 
-On Windows, you may need to install windows-curses (required for getting button presses in the psychoacoustics classes):
+Other optional requirements can be installed by telling pip which extras you want:
 
-```pip install windows-curses```
+```pip install slab[name_of_extra]```
+
+The options for `name_of_extra` are:
+- `windows`: if you are running Windows - this will install windows-curses for you, which is required for getting button presses in the psychoacoustics classes,
+- `hrtf`: if you want to use spatial stimuli with the `Binaural` and `HRTF` classes,
+- `testing`: (for developers) if you want to run the unit tests for slab, and
+- `docs`: (for developers) if you want to build the documentation locally.
+
+You can combine these options: `pip install slab[windows, hrtf]` if you are on Windows and use spatial sounds.
 
 Detailed installation instructions can be found [here](https://slab.readthedocs.io/en/latest/index.html#installation).
 
