### 2021-05-20 #3284 @arfon
> Alright, it's a pass for me! Really great job on this lib @DrMarc (and your acolyte @OleBialas). I'm be telling my labmates working on psychoacoustics about it :)

:zap: thanks @hadware!
