### 2021-05-22 64f2b98
_calibration_intensity behaves like _default_samplerate, _default_level
