+++ slab/sound.py
@@ -37,22 +37,18 @@ if _system == 'Windows':
 import slab.signal
 from slab.signal import Signal
 from slab.filter import Filter
-from slab import data_path
 
-# get a temporary directory for writing intermediate files
-_tmpdir = pathlib.Path(tempfile.gettempdir())
-
-try:  # try getting a previously set calibration intensity from file
-    _calibration_intensity = numpy.load(data_path(allow_download=False) + 'calibration_intensity.npy')
-except FileNotFoundError:
-    _calibration_intensity = 0  #: Difference between rms intensity and measured output intensity in dB
-
+_tmpdir = pathlib.Path(tempfile.gettempdir())  # get a temporary directory for writing intermediate files
+_calibration_intensity = 0  # difference between rms intensity and measured output intensity in dB
 _default_level = 70  # the default level for generated Sounds in dB
-
 
 def set_default_level(level):
     global _default_level
     _default_level = level
+
+def set_calibration_intensity(intensity):
+    global _calibration_intensity
+    _calibration_intensity = intensity
 
 
 class Sound(Signal):
@@ -200,7 +196,6 @@ class Sound(Signal):
 
     @staticmethod
     def harmoniccomplex(f0=500, duration=1., amplitude=0, phase=0, samplerate=None, n_channels=1):
-        # TODO: in tone() the phase argument refers to the channels, here it refers to the harmonics --> rename?
         """
         Generate a harmonic complex tone composed of pure tones at integer multiples of the fundamental frequency.
 
@@ -623,6 +618,10 @@ class Sound(Signal):
         Returns:
             (slab.Sound): the input sounds combined in a single object.
         """
+        samplerate = sounds[0].samplerate
+        for sound in sounds:
+            if sound.samplerate != samplerate:
+                raise ValueError('All sounds must have the same sample rate.')
         samplerate = sounds[0].samplerate
         for sound in sounds:
             if sound.samplerate != samplerate:
@@ -1267,32 +1266,35 @@ class Sound(Signal):
         return numpy.array(samplepoints) / self.samplerate  # convert to array of time points
 
 
-def calibrate(intensity=None, make_permanent=False):
+def calibrate(sound=None):
     """
-    Calibrate the presentation intensity of a setup. The calibration intensity determines the relationship
-    between the amplitude of a sound and it's level.
+    Calibrate the presentation intensity of a setup by playing a sound of known intensity (default is a 1 kHz pure tone
+    at 70 dB, played for 5 s). Measure the playback intensity of this sound and enter it when requested. For easy and
+    accurate measurement, the sound should be several seconds long. Use an appropriate acoustic coupler (artificial ear)
+    when calibrating headphones.
+    The return value is the difference between the measured playback intensity and the sound's RMS level (as calculated
+    by the `level` attibute). For example, if the `level` of the sound is 70 dB and you record 60 dB, the returned value
+    will be -10. If you pass this value to `slab.set_calibration_intensity`, then the `level` attribute of all sounds
+    will approximate their playback intensity (if using the same headphones or loudspeakers used in the calibration).
+
+    To increase the accuracy of the calibration for your experimental stimuli, pass a sound with a similar spectrum.
+    For instance, if your stimuli are wide band pink noises, then you may want to use a pink noise for calibration.
+    The `level` of the noise should be high, but not cause clipping.
 
     Arguments:
-        intensity (int | float | None): the difference between the actual intensity of the output and the intensity
-            indicated by the sound's `level` attribute (For example: If `level` is 80 dB and you record 60 dB, the
-            value of `intensity` should be -20). If None, a 1 kHz tone is played for five seconds after which you will
-             be prompted to input the measured intensity.
[…]
