### 2021-05-22 8e13a39
added slab.set_calibration_intensity
