+++ slab/__init__.py
@@ -9,6 +9,6 @@ sys.path.append('..\\')
 from slab.hrtf import HRTF
 from slab.psychoacoustics import *
 from slab.binaural import Binaural
-from slab.sound import Sound, set_default_level, set_calibration_intensity
+from slab.sound import Sound, set_default_level, set_calibration_intensity, calibrate
 from slab.signal import Signal, set_default_samplerate
 from slab.filter import Filter
