### 2021-05-22 76e598d
fixes #35, clarified calibration and default level
