+++ slab/__init__.py
@@ -1,6 +1,5 @@
-import pathlib
-import os
 import sys
+import pathlib
 
 __version__ = '0.9.0'
 
