+++ setup.py
@@ -28,4 +28,6 @@ setup(name='slab',
                       'docs': ['sphinx', 'sphinx-rtd-theme'],
                       'hrtf': ['h5netcdf']},
       packages=find_packages(),
-      zip_safe=True)
+      package_data={'slab': ['data/mit_kemar_normal_pinna.bz2']},
+      include_package_data=True,
+      zip_safe=False)
