+++ MANIFEST.in
@@ -1,9 +1,5 @@
 include README.md
+include slab/data/mit_kemar_normal_pinna.bz2
 graft slab/experiments/
 prune slab/experiments/__pycache__/
 prune slab/experiments/.DS_Store
-prune slab/data/mit_kemar_normal_pinna.sofa
-prune slab/data/KEMAR_interaural_level_spectrum.npy
-prune slab/data/aachen/
-prune slab/data/ari/
-prune slab/data/cipic/

[…]
