### 2021-05-26 2cf6d96
describe changed level spectrum handling
