### 2021-05-26 6fb7900
add level argument to generator methods
