+++ slab/binaural.py
@@ -412,52 +412,47 @@ class Binaural(Sound):
         return out.resample(samplerate=original_samplerate)
 
     @staticmethod
-    def whitenoise(duration=1.0, kind='diotic', samplerate=None):
+    def whitenoise(kind='diotic', **kwargs):
         """
         Generate binaural white noise. `kind`='diotic' produces the same noise samples in both channels,
         `kind`='dichotic' produces uncorrelated noise. The rest is identical to `slab.Sound.whitenoise`.
         """
-        out = Binaural(Sound.whitenoise(duration=duration, n_channels=2, samplerate=samplerate))
+        out = Binaural(Sound.whitenoise(**kwargs))
         if kind == 'diotic':
             out.left = out.right
         return out
 
     @staticmethod
-    def pinknoise(duration=1.0, kind='diotic', samplerate=None):
+    def pinknoise(kind='diotic', **kwargs):
         """
         Generate binaural pink noise. `kind`='diotic' produces the same noise samples in both channels,
         `kind`='dichotic' produces uncorrelated noise. The rest is identical to `slab.Sound.pinknoise`.
         """
-        return Binaural.powerlawnoise(
-                duration=duration, alpha=1, samplerate=samplerate)
+        return Binaural.powerlawnoise(alpha=1, **kwargs)
 
     @staticmethod
-    def powerlawnoise(duration=1.0, alpha=1, kind='diotic', samplerate=None):
+    def powerlawnoise(kind='diotic', **kwargs):
         """
         Generate binaural power law noise. `kind`='diotic' produces the same noise samples in both channels,
         `kind`='dichotic' produces uncorrelated noise. The rest is identical to `slab.Sound.powerlawnoise`.
         """
         if kind == 'dichotic':
-            out = Binaural(Sound.powerlawnoise(
-                duration=duration, alpha=alpha, samplerate=samplerate, n_channels=2))
-            out.left = Sound.powerlawnoise(
-                duration=duration, alpha=alpha, samplerate=samplerate, n_channels=1)
+            out = Binaural(Sound.powerlawnoise(n_channels=2, **kwargs))
+            out.left = Sound.powerlawnoise(n_channels=1, **kwargs)
         elif kind == 'diotic':
-            out = Binaural(Sound.powerlawnoise(
-                duration=duration, alpha=alpha, samplerate=samplerate, n_channels=2))
+            out = Binaural(Sound.powerlawnoise(n_channels=2, **kwargs))
             out.left = out.right
         else:
             raise ValueError("kind must be 'dichotic' or 'diotic'.")
         return out
 
     @staticmethod
-    def irn(frequency=100, gain=1, n_iter=4, duration=1.0, kind='diotic', samplerate=None):
+    def irn(kind='diotic', **kwargs):
         """
         Generate iterated ripple noise (IRN). `kind`='diotic' produces the same noise samples in both channels,
         `kind`='dichotic' produces uncorrelated noise. The rest is identical to `slab.Sound.irn`.
         """
-        out = Binaural(Sound.irn(frequency=frequency, gain=gain, n_iter=n_iter, duration=duration,
-            samplerate=samplerate, n_channels=2))
+        out = Binaural(Sound.irn(n_channels=2, **kwargs))
         if kind == 'diotic':
             out.left = out.right
         else:
@@ -465,21 +460,19 @@ class Binaural(Sound):
         return out
 
     @staticmethod
-    def tone(frequency=500, duration=1., phase=0, samplerate=None):
+    def tone(**kwargs):
         """ Identical to slab.Sound.tone, but with two channels. """
[…]
