+++ slab/sound.py
@@ -159,7 +159,7 @@ class Sound(Signal):
         return Sound(data, samplerate=samplerate)
 
     @staticmethod
-    def tone(frequency=500, duration=1., phase=0, samplerate=None, n_channels=1):
+    def tone(frequency=500, duration=1., phase=0, samplerate=None, level=None, n_channels=1):
         """
         Generate a pure tone.
 
@@ -170,6 +170,8 @@ class Sound(Signal):
             phase (int | float | list): phase of the sinusoid, defaults to 0. Given a list of length `n_channels`, one
                 element of the list is used as phase for each channel.
             samplerate (int | None): the samplerate of the sound. If None, use the default samplerate.
+            level (None | int | float | list): the sounds level in decibel. For a multichannel sound, a list of values
+                can be provided to set the level of each channel individually. If None, the level is set to the default
             n_channels (int): number of channels, defaults to one.
         Returns:
             (slab.Sound): the tone generated from the parameters.
@@ -191,11 +193,11 @@ class Sound(Signal):
         t.shape = (t.size, 1)  # ensures C-order
         x = numpy.sin(phase + 2 * numpy.pi * frequency * numpy.tile(t, (1, n_channels)))
         out = Sound(x, samplerate)
-        out.level = _default_level
+        out.level = level
         return out
 
     @staticmethod
-    def harmoniccomplex(f0=500, duration=1., amplitude=0, phase=0, samplerate=None, n_channels=1):
+    def harmoniccomplex(f0=500, duration=1., amplitude=0, phase=0, samplerate=None, level=None, n_channels=1):
         """
         Generate a harmonic complex tone composed of pure tones at integer multiples of the fundamental frequency.
 
@@ -212,6 +214,8 @@ class Sound(Signal):
                 harmonic. Given a string, its value must be schroeder', in which case the harmonics are in
                 Schroeder phase, producing a complex tone with minimal peak-to-peak amplitudes (Schroeder 1970).
             samplerate (int | None): the samplerate of the sound. If None, use the default samplerate.
+            level (None | int | float | list): the sounds level in decibel. For a multichannel sound, a list of values
+                can be provided to set the level of each channel individually. If None, the level is set to the default
             n_channels (int): number of channels, defaults to one.
         Returns:
             (slab.Sound): the harmonic complex generated from the parameters.
@@ -246,17 +250,19 @@ class Sound(Signal):
                              phase=phases[i], samplerate=samplerate, n_channels=n_channels)
             tmp.level = lvl + amplitudes[i]
             out += tmp
-        out.level = _default_level
+        out.level = level
         return out
 
     @staticmethod
-    def whitenoise(duration=1.0, samplerate=None, n_channels=1):
+    def whitenoise(duration=1.0, samplerate=None, level=None, n_channels=1):
         """
         Generate white noise.
 
         Arguments:
             duration (float | int): duration of the sound in seconds (given a float) or in samples (given an int).
             samplerate (int | None): the samplerate of the sound. If None, use the default samplerate.
+            level (None | int | float | list): the sounds level in decibel. For a multichannel sound, a list of values
+                can be provided to set the level of each channel individually.
             n_channels (int): number of channels, defaults to one. If channels > 1, several channels of uncorrelated
                 noise are generated.
         Returns:
@@ -270,11 +276,11 @@ class Sound(Signal):
         duration = Sound.in_samples(duration, samplerate)
         x = numpy.random.randn(duration, n_channels)
         out = Sound(x, samplerate)
-        out.level = _default_level
+        out.level = level
         return out
 
     @staticmethod
-    def powerlawnoise(duration=1.0, alpha=1, samplerate=None, n_channels=1):
[…]
