### 2021-05-27 #3284 @DrMarc
@arfon - Ole and I would like to indicate that each of us contributed 50% to the work (sometimes called shared first authorship). Is there a preferred way to do that at JOSS?
