### 2021-05-27 #3284 @arfon
See the example docs on how to do this (https://joss.readthedocs.io/en/latest/submitting.html#example-paper-and-bibliography). Basically add `^[co-first author]` after each of the names.
