### 2021-05-28 f625466
remove all plot kwargs and add kwargs to spectrogram
