### 2021-05-28 21a81dc
work on hrtf interpolation, some clean-up
