### 2021-05-28 b5b11cf
kwargs & bugs (#40)
* exclude dynamicripple from pytest

* test workflow

* add flake8 config file

* exclude soundfile from flake8

* exclude soundcard from flake8

* ignore soundcard for pytest

* ignore soundfile for pytest

* fix bu in test routine

* don't check sound level

* add testing for filters

* change filter documentation

* import re

* change filters to filter in documentation

* change f_lower to low_cutoff

* remove brownnoise

* ignore build

* adding filter doc

* ignore build

* ignore build

* ignore build

* removing _build folder

* convert tabs to spaces

* debug calib
[…]
