### 2021-06-01 38f7be5
increment version to reflect changes due to JOSS review so far
