+++ slab/__init__.py
@@ -1,7 +1,7 @@
 import sys
 import pathlib
 
-__version__ = '0.9.0'
+__version__ = '0.9.5'
 
 sys.path.append('..\\')
 
