### 2021-06-09 1ff4521
added channel colour in plot_samples
