+++ slab/hrtf.py
@@ -16,6 +16,7 @@ except ImportError:
     matplotlib, plt = False, False
 try:
     from mpl_toolkits.mplot3d import Axes3D
+    from mpl_toolkits.mplot3d.art3d import Poly3DCollection
     from mpl_toolkits.axes_grid1 import make_axes_locatable
 except ImportError:
     Axes3D = False
@@ -26,9 +27,11 @@ except ImportError:
     h5netcdf = False
 try:
     import scipy.signal
+    import scipy.spatial
 except ImportError:
     scipy = False
 from slab.signal import Signal
+from slab.sound import Sound
 from slab.filter import Filter
 
 _kemar = None
@@ -458,24 +461,24 @@ class HRTF:
             tfs[:, idx] = jwd.flatten()
         return tfs
 
-    def interpolate(self, azimuth=0, elevation=0, method='nearest'):
+    def interpolate(self, azimuth=0, elevation=0, method='nearest', plot_tri=False):
         """
         Interpolate a filter at a given azimuth and elevation from the neighboring HRTFs. A weighted average of the
-        n closest HRTFs in the set is computed, after removing group-delay differences between filters. The interaural
-        time difference at the requested directio is interpolated separately and added to the interpolated filter.
-        The resulting filter values vary smoothly with changes in azimuth and elevation.
-        The fidelity of the interpolated filter decreases with increasing distance of the closest sources and should
-        only be regarded as appropriate approximation when the contributing filters are less than 20˚ away.
+        3 closest HRTFs in the set is computed in the spectral domain with barycentric weights. The resulting filter
+        values vary smoothly with changes in azimuth and elevation. The fidelity of the interpolated filter decreases
+        with increasing distance of the closest sources and should only be regarded as appropriate approximation when
+        the contributing filters are less than 20˚ away.
 
         Arguments:
             azimuth (float): the azimuth component of the direction of the interpolated filter
             elevation (float): the elevation component of the direction of the interpolated filter
-            method (str): interpolation method, 'nearest' returns the filter of the nearest direction.
+            method (str): interpolation method, 'nearest' returns the filter of the nearest direction. Any other string
+                returns a barycentric interpolation.
+            plot_tri (bool): plot the triangulation of source positions used of interpolation. Useful for checking
+                for areas where the interpolation may not be accurate (look for irregular or elongated triangles).
         Returns:
             (slab.Filter): a 2-channel Filter, interpolated from the neighboring filters in the set
         """
-        if method != 'nearest':
-            raise NotImplementedError('Only nearest-neightbour interpolation is currently supported.')
         # spherical to cartesian
         coords = self.cartesian_source_locations()
         r = self.sources[:, 2].mean()
@@ -485,6 +488,76 @@ class HRTF:
         if method == 'nearest':
             idx_nearest = numpy.argmin(distances)
             return self.data[idx_nearest]
+        # triangulate source positions into triangles
+        if not scipy:
+            raise ImportError('Need scipy.spatial for barycentric interpolation.')
+        tri = scipy.spatial.ConvexHull(coords)
+        if plot_tri:
+            ax = plt.subplot(projection='3d')
+            for simplex in tri.points[tri.simplices]:
+                polygon = Poly3DCollection([simplex])
+                polygon.set_color(numpy.random.rand(3))
+                ax.add_collection3d(polygon)
+                mins = coords.min(axis=0)
[…]
