### 2021-06-11 b8c266c
refactor init to avoid 'object has samplerate' warning
