### 2021-06-12 87c893d
'left' and 'right' properties now return Sound object, not Binaural
