### 2021-06-15 DrMarc/slab#41 @sneakers-the-rat
Testing/Maintaining Binaural class
(optional comment/suggestion)

The Binaural class does what it promises, but as written it seems a little hard to maintain -- sound generation methods from `Sound` have to be overwritten with a call to `Binaural(Sound.method(**kwargs))` . I think you would do yourselves a favor by doing one or two things:

* making an explicit declaration about which methods of the `Sound` class are sound generation methods (looks like they are all `@classmethod`s, so you could make a decorator that wraps `@classmethod` like `@sound` or something that adds to a class-level list?), and then write a test that confirms that for each of the methods in `Sound`, `Binaural` produces the expected behavior

or

* embed some logic in `Binaural` that doubles channels automatically unless the generation method has been overridden... i imagine this could be done by making `self.data` a property with a getter/setter, or else some wrapper around the sound generation methods that does it.

i am just imagining those getting out of sync becoming a headache, hopefully useful, but up to you :)
