### 2021-06-15 DrMarc/slab#42 @sneakers-the-rat
Psychoacoustics Section
Some (optional, hopefully useful) comments on the docs/etc. of the psychoacoustics section. Grouped them all here to not open a zillion related issues.

first of all i like the trialsequence object a lot, nice work. there's a lot of functionality packed into this module, i think it would be worth splitting up/expanding the docs here because this is basically a mini experimental framework that i think gets shortchanged in the existing docs/paper relative to how useful/# features it has as i unpacked the docs a bit.

- the api docs for the trialsequence class could use a little extra description of the format of the `.json` file that is saved with `save` (it's implicit in the structure of the object, but i think it's worth describing what the object expects/will produce).
- the `psychoacoustics.key()` function seems like it could use the ability to present some prompt text, would that be too challenging? in the example task i didn't read carefully and left left and right arrows and didn't know that i was getting anything wrong, so might be worth it to model in the docs something that corrects users if they give an unexpected input, even if that's pedantic.
- transitions are documented in the api docs, but could use a brief sentence explaining what the rows/columns are. the docs are sort of unclear to me -- i'm not sure if the values in the table represent the number of times that condition will be played in sequence (i think as described in the docs) in which case is it deterministic? etc. or the transition probability (maybe normalized to 1)?
- the docs get a little blippy near the section on plotting staircases, sort of lose me on implementation of eg. `input_method`, `key_press_event` -- I tried to run the staircase example but i didn't know what key i was supposed to press. It seems like it expects a key index of 0, not sure what that is!
- `input_method` is documented most fully in the `recording responses` , but maybe that should be called 'input method' or 'hardware input' or something like that, at least that's what i was expecting/looking for
- `Key` is referred to as `Key` class, when i believe it is now a function in the text of the docs in recording responses
- I like the `ResultsFile` class, but i couldn't quite figure out what to expect from it from the docs. might be worth putting an example in the api docs
- `ResultsFile` says 'safely' and i would typically interpret that to mean 'thread safe' -- a `threading.Lock` around the `open` call in `write` would give you that, and resolve the ambiguity in the docs. as is the safety is sort of implicit, but there could be problems if eg. used in a multithreaded context which i would expect to work given the description of 'safe.'
- the use of a custom .txt format and namedtuples in `load_config`  is confusing to me given the use of .json elsewhere, but it seems to only be mentioned in the bottom of the psychoacoustics section so might not be all that important to make uniform?
