### 2021-06-15 DrMarc/slab#43 @sneakers-the-rat
Filter Section
Going to group my optional comments here (as i don't really anticipate having any serious issues with the rest of the package)

- seems like `Filter.band` could take `**kwargs` and pass them to the scipy filter-generating function, but I didn't look to closely at how this was implemented to see how hard that would be.
- the axes are mislabeled here for `tf()` - looks like x and y labels should be swapped:

- The behavior of multiple filters is the opposite of what i would expect. I think this is probably "just a me thing" but i think it points at a place where the docs could walk all the "just me"s in the world along :) -- I would expect multiple filters to operate as a compound filter, so i could eg. make a double bandpass filter, so applying it to a multichannel sound should apply both filters to both channels. I think the intended way of doing this is to provide a list of frequency to a single filter object, right? Currently it applies each filter to the corresponding channel (as described correctly in the docs). I would want some option to be able to use it the former way, so an arg like `each=True` in `apply` would resolve the ambiguity to me. The next section describes how to re-combine the bands, but creating a multichannel sound, applying a filter to each of its channels, and then combining each of its subbands is less intuitive to me than being able to choose whether to apply a multichannel filter as a compound/sequential/what-have-you filter or to split it across the channels. I see the utility of the splitting method, as demonstrated with modifying the bands independently before recombining, so definitely not saying get rid of it, just a little sugar to sweeten the method :).
- the noise vocoded speech example isn't reproducible -- `signal` isn't declared, seems like it would be good to include a quick example .wav in the docs or else explain how to get `signal`. I see that it's the contents of the `vocode` method, but that seems odd given the rest of the docs are examples. ps i think you mean "omitted" rather than "redacted" ;)
- I think that the example of equalizing at the end here is adequate for our purposes here, but i would strongly suggest giving some further example of how one might measure the impulse response of their audio setup. It could be as simple as linking to some youtube video, or the wiki page for "impulse response" but if I were a naive reader who had never measured an impulse response I wouldn't really know what to do here.
- the "equalization" header is repeated twice.
