### 2021-06-15 DrMarc/slab#44 @sneakers-the-rat
Docs out of sync - `data_path()`
The hrtf docs (https://github.com/DrMarc/slab/blob/5a798d5f51eeb7c1f5ac8bbab5c336a5d62d5415/docs/hrtf.rst) currently read:

```
hrtf = slab.HRTF(data=slab.data_path()+'mit_kemar_normal_pinna.sofa')
```

but it looks like `data_path()` was removed and I get:

```
AttributeError: module 'slab' has no attribute 'data_path'
```
