### 2021-06-15 DrMarc/slab#45 @sneakers-the-rat
Docs out of sync - Filter.tf(label=...)
Docs currently include an example:

```
from matplotlib import pyplot as plt
hrtf = slab.HRTF(data=slab.data_path()+'mit_kemar_normal_pinna.sofa')
fig, ax = plt.subplots(1)
idx = 10
source = hrtf.sources[idx]  # the source's azimuth, elevation and distance
filt = hrtf.data[idx] # the corresponding filter
fig.suptitle(f"source at azimuth {source[0].round(2)} and elevation {source[1]}")
filt.channel(0).tf(axis=ax, show=False, label="left ear")
filt.channel(1).tf(axis=ax, show=False, label="right ear")
plt.legend()
plt.show()
```

but `tf` has no property `label`:

```
TypeError: tf() got an unexpected keyword argument 'label'
```
