### 2021-06-15 #3284 @sneakers-the-rat
whup whup @DrMarc and @OleBialas this is a big pass from me!

I left some last (optional) comments in the final issues that might be helpful:
- https://github.com/DrMarc/slab/issues/41
- https://github.com/DrMarc/slab/issues/42
- https://github.com/DrMarc/slab/issues/43
- https://github.com/DrMarc/slab/issues/44
- https://github.com/DrMarc/slab/issues/45

really excellent work! love this package, after i got a hang on the design i really enjoyed using it while testing it out. will absolutely recommend this to anyone looking for a simple to use but powerful psychoacoustics package!
