### 2021-06-15 #3284 @DrMarc
Hi @sneakers-the-rat,
Many thanks for the details comments and bug-hunting! This is very valuable. We will correct and expand the docs, as you suggested, and take care of the other issues. Some of these issues are intuitive to us, but may not be to others. The multi-channel filter is an example (#43). We work with multi-channel sound (48-channel speaker array in the lab), and our typical use case is applying each filter channel to each sound channel. But I can see that the sequential application may be useful in other scenarios.
Thanks again for your work on this!
