### 2021-06-15 DrMarc/slab#43 @OleBialas
We chose the behavior of the multi-channel filter on the multi-channel sound based on our use cases but you are right in that it may be counterintuitive to outsiders.
Pointing to additional info regarding the recording of impulse responses is a great idea!
