### 2021-06-15 DrMarc/slab#41 @OleBialas
I like the idea of a decorator for the sound generating methods better than the "automatic doubling"
