### 2021-06-15 022efbf
corrected 'bs' filter kind (removed 'notch')
