### 2021-06-16 DrMarc/slab#43 @sneakers-the-rat
Yes i think both the "compound filter" (have no idea if that is even a term) and multichannel modes are both valid -- didn't mean to muddy the waters, just giving an outside perspective (i usually only work with one or two channel audio!), so no worries if you want to keep it as is no changes, i certainly won't be offended :)
