### 2021-06-19 DrMarc/slab#41 @DrMarc
Checking this in the tests is a great idea. We can do that without adding any code to the Sound or Binaural classes by using inspection to get all methods with a 'n_channels' parameter from the Sound class and asserting that they return a 2chan object when called from Binaural, like so:
methods = [attribute for attribute in dir(slab.Sound) if callable(getattr(slab.Sound, attribute)) and attribute.startswith('__') is False]
for method in methods:
    func = getattr(slab.Sound, 'whitenoise')
    args, _, _ = inspect.getargs(func.__code__)
    if 'n_channels' in args:
        # same method in Binaural returns 2-ch object?
        assert(getattr(slab.Binaural, 'whitenoise')().n_channels == 2)

I'll add this test in the next commit.
