### 2021-06-19 DrMarc/slab#43 @OleBialas
Ok so after thinking about it for a while we came to the conclusion that the clearest way is to not add any arguments to the apply function. Instead one can apply a bank of filters to one sound in a for loop, like:
```
    for i in range(filter_bank.n_channels):
        sound = filter_bank.channel(i).apply(sound)
```
I will add an example of this to the documentation
