### 2021-06-19 16c9071
add example for sequential application of filters
