+++ docs/filter.rst
@@ -37,9 +37,12 @@ Inspecting the waveform of the sound (using the :meth:`~slab.Sound.waveform` met
 
 Filter banks
 ------------
-A :class:`Filter` objects can hold multiple channels, just like a :class:`Sound` object. The :meth:`.cos_filterbank` and :meth:`.equalizing_filterbank` methods
-construct multi-channel filter banks for you, but you can also initialize a :class:`Filter` object with a list of filters, just as you can make a multi-channel
-sound by initializing a :class:`Sound` object with a list of sounds (both classes inherit this feature from the parent :class:`Signal` class):
+A :class:`Filter` objects can hold multiple channels, just like a :class:`Sound` object. In the following, we will refer
+to filters with multiple channels as filter banks. You can create filter banks the same way you create multi-channel
+sound (since the :class:`Filter` and :class:`Sound` class both inherit from the parent :class:`Signal` class).
+When you apply a filter bank to a single sound, each filter will be applied to a separate copy of the sound and the
+:func:`apply` function will return a sound with a number of channels equal to the number of filters in the bank.
+This way you can create, for example, a series of sounds with different frequency bands:
 
 .. plot::
     :include-source:
@@ -61,33 +64,12 @@ sound by initializing a :class:`Sound` object with a list of sounds (both classe
 .. plot::
     :context: close-figs
 
-If the a one-channel filter is applied to a multi-channel signal, the filter will be applied to each
-channel individually. This can be used, for example, to easily pre-process a set of recordingss (where
-every recordings is represented by a channel in the :class:`slab.Sound` object). If a multi-channel filter
-is applied to a multi-channel signal with the same number of channels each filter channel is applied to
-the corresponding signal channel. This is useful for e.g. equalization of a set of loudspeakers
-
-Equalization
-------------
-
-In Psychoacoustic experiments, we are often interested in the effect of a specific feature. One could,
-for example, take the bandpass filtered sounds from the example above and investigate how well listeners
-can discriminate them from a noisy background - a typical cocktail-party task. However, if the transfer
-function of the loudspeakers or headphones used in the experiment is not flat, the findings will be biased.
-Imagine that the headphones used were bad at transmitting frequencies below 1000 Hz. This would make a sound
-with center frequency of 550 Hz harder to detect than one with a center frequency of 1550 Hz. We can prevent
-this by inverting the headphones transfer function and pre-filter our stimuli with this inverse. The inverse transfer function and the actual transfer function will then cancel each other out and result in an equalized sound.
-
-.. plot::
-    :include-source:
-    :context: close-figs
-
-    for freq in range(200, 3000, 300):
-        filters.append(slab.Filter.band(frequency=(freq, freq+200), kind='bp'))
-    fbank = slab.Filter(filters)
-    fbank.tf()
-
-If this multi-channel filter is applied to a one-channel signal, each filter channel is applied separately and the resulting signal has the same number of channels as the filter (subbands). You can modify these subbands and re-combine them using the :meth:`.combine_subbands` method. An example of this process is the vocoder implementation in the :class:`Sound` class, which uses these features of the :class:`Filter` class. The multi-channel filter is generated with :meth:`.cos_filterbank`, which produces cosine-shaped filters that divide the sound into small frequency bands which are spaced in a way that mimics the filters of the human auditory periphery (`equivalent rectangular bandwidth, ERB <https://en.wikipedia.org/wiki/Equivalent_rectangular_bandwidth>`_). Here is an example of the transfer functions of this filter bank:
+The channels, or subbands, of the filtered sound can be modified and re-combine with the :meth:`.combine_subbands`
+method. An example of this process is the vocoder implementation in the :class:`Sound` class, which uses these features
[…]
