### 2021-06-19 DrMarc/slab#43 @OleBialas
I could not find a proper source for a guide on recording transfer functions. The ones I found where either way to technical for our audience or instructions for some commercial software. I made some slight modifications in the docs and i think the info is detailed enough for our purposes
