### 2021-06-19 DrMarc/slab#42 @OleBialas
You are right - the docs talk about "transition probabilities" while they are actually counts - I clarified it.
