### 2021-06-19 DrMarc/slab#42 @OleBialas
About writing "safely" in the ResultsFile: we were thinking of "safely" in the sense that you don't loose any data if the Experiment crashes because you create a new entry after every trial - I understand the confusion and will try to make it clearer in the docs.
As for the threading lock - as we don't anticipate any multi threading use case I would rather not. It might be confusing to users who never grappled with the concept
