### 2021-06-20 7361b39
added option to print a message to terminal when getting a button press
