### 2021-06-20 4e298f1
improved description of key press acquisition
