### 2021-06-20 5c0b837
can now specify more than one correct key in present_tone_trial
