### 2021-06-21 DrMarc/slab#43 @DrMarc
Thanks for the suggestion. filter.band uses scipy.signal.firwin2 under the hood, and that function does not have any kwargs that we are not already using, i.e. there is nothing more to set. However, there are some situations in which one may want to set things like the filter order directly and use, for instance, a Butterworth filter. IIR filters are currently not supported by the Filter, but a one-off method in the Sound class (or extension of Sound.filter) would work. We'll consider that.
We have addressed the other points in the last commits.
