### 2021-06-21 570c9bd
improved explanation of load_config format
