### 2021-06-21 e58e7fc
typos in trail_sequence description
