### 2021-06-21 cb28b02
minor improvement of load_config description
