### 2021-06-21 DrMarc/slab#42 @DrMarc
Alright - we have covered the above suggestions in the last few commits.
