### 2021-06-22 af33e4e
added dynamic_tone generation method
