### 2021-06-22 c809820
added description of HRTF interpolation
