### 2021-06-23 DrMarc/slab#43 @sneakers-the-rat
thanks for considering these, still getting used to the structure of this new style of review, i just meant these as food for thought to take or leave -- it's your package and your design, so whatever you think is best by all means do it :).

the syntax for applying compound filters is a little confusing to me, but if you've got an example, it's the canonical way to do it with your package, and the ambiguity is articulated away i think that sounds lovely :)

for measuring an impulse response I think the typical way is to play a pulse (whether that's a literal 0 - 1 single bit pulse or a longer white-noise pulse) and the following should be the impulse response in the time domain, the other way I know of is to play a sine sweep or some other flat sound, measure the emitted sound, and take the difference in the fft domain. but yeah maybe a little out of scope, if you can't find something to link to then i think just a link to the wiki article for impulse response would be fine.

and yes my comments re `**kwargs`-ing things mostly are oriented towards keeping things maintainable for y'all so you don't end up with a zillion arguments in the wrapper functions/methods, but since there is a reasonable amount of additional logic in your functions i think that's less relevant.
