### 2021-06-23 DrMarc/slab#41 @sneakers-the-rat
well i am always a sucker for introspection for things that are internally-facing rather than part of an API, and i just want to say i like this solution^
