### 2021-06-23 2a45651
ITDs are interpolated correctly by HRTF.interpolate; hrtf[idx] now works
