+++ slab/hrtf.py
@@ -44,10 +44,10 @@ class HRTF:
         data (str | Filter | numpy.ndarray): Typically, this is the path to a file in the .sofa format.
             The file is then loaded and the data of each source for which the transfer function was recorded is stored
             as a Filter object in the `data` attribute. Instead of a file name, the data can be passed directly as
-            Filter or numpy array (not recommended). Given a `Filter`, every filter channel in the instance is taken as
-            a sound source. Given an 3-dimensional array, the first dimension represents the sources, the second the
-            number of taps per filter and the last the number of filter channels per filter (should be always 2 for
-            left and right ear).
+            Filter or numpy array. Given a `Filter`, every filter channel in the instance is taken as a source (this
+            does not result in a typical HRTF object and is only intended for equalization filter banks). Given an 3D
+            array, the first dimension represents the sources, the second the number of taps per filter and the last the
+            number of filter channels per filter (should be always 2, for left and right ear).
         samplerate (None | float): rate at which the data was acquired, only relevant when not loading from .sofa file
         sources (None | array): positions of the recorded sources, only relevant when not loading from .sofa file
         listener (None | list | dict): position of the listener, only relevant when not loading from .sofa file
@@ -105,24 +105,33 @@ class HRTF:
                 self.listener = [0, 0, 0]
             else:
                 self.listener = listener
-        else:
+        elif isinstance(data, numpy.ndarray):  # should have shape (ind x taps x ear), 2 x n_taps filter (left right)
+            if samplerate is None:
+                raise ValueError('Must specify samplerate when initialising HRTF from an array.')
             self.samplerate = samplerate
             self.data = []
             for idx in range(data.shape[0]):
-                # (ind x taps x ear), 2 x n_taps filter (left right)
                 self.data.append(Filter(data[idx, :, :].T, self.samplerate))
             self.sources = sources
             if listener is None:
                 self.listener = [0, 0, 0]
             else:
                 self.listener = listener
+        else:
+            raise ValueError(f'Unsupported data type: {type(data)}')
 
     def __repr__(self):
         return f'{type(self)} (\n{repr(self.data)} \n{repr(self.samplerate)})'
 
     def __str__(self):
         return f'{type(self)} sources {self.n_sources}, elevations {self.n_elevations},' \
-               f'samples {self.data[0].n_samples}, samplerate {self.samplerate}'
+               f'samples {self[0].n_samples}, samplerate {self.samplerate}'
+
+    def __getitem__(self, key):
+        return self.data.__getitem__(key)
+
+    def __setitem__(self, key, value):
+        return self.data.__setitem__(key, value)
 
     # Static methods (used in __init__)
     @staticmethod
@@ -236,19 +245,20 @@ class HRTF:
         Returns:
             (slab.Binaural): a spatialized copy of `sound`.
         """
+        from slab.binaural import Binaural  # inporting here to avoid circular import at top of class
         if (sound.samplerate != self.samplerate) and (not allow_resampling):
             raise ValueError('Filter and sound must have same sampling rates.')
         original_rate = sound.samplerate
         sound = sound.resample(self.samplerate) # does nothing if samplerates are the same
-        left = scipy.signal.fftconvolve(sound[:, 0], self.data[source][:, 0])
+        left = scipy.signal.fftconvolve(sound[:, 0], self[source][:, 0])
         if sound.n_channels == 1:
[…]
