### 2021-06-23 b3ca35d
corrected HRTF.interpolate description
