### 2021-06-24 113ffef
corrected test of HRTF generation from array (must provide samplerate)
