### 2021-06-24 ef9a4e7
HRTF.interpolate returns HRTF object instead of filter (correct .apply)
