### 2021-06-24 b6264a9
version increment (minor fixes of JOSS release)
