+++ slab/__init__.py
@@ -1,7 +1,7 @@
 import sys
 import pathlib
 
-__version__ = '1.0.0'
+__version__ = '1.0.1'
 
 sys.path.append('..\\')
 
